requirements.txt
pandas
pymongo
mysql-connector-python
python-dotenv
llama-cpp-python
modelcontextprotocol
